Removing Bower support
In favor of Browserify and npm.

Dependencies can now be installed with npm and `require`d or
`import`ed by Browserify and bundled with the final
application.

That removes the need to bundle all exterrnal dependencies
provided by Bower, as a separate script file.

Phaser is not a Node.js-compatible library, though, requiring
a dedicated task to copy a chosen build file, and still
provide it as a global library.