Removed some fields from `package.json`.
As this project is never meant to be published as a package
under npm, having the `name`, `version`, `description` and
`license` fields declared are not a requirement. It is very
unlikely that such a project would get published to npm,
though we completely void that possibility by removing these
required fields.

Also, removed the `browserify` field, as Browserify have its
own dedicated configuration module.