Deleted an old screen orientation warning graphic.
Last remains from the old sample game that used to be
distributed in earlier versions.

As of now, there is no routine dealing with screen orientation
in the sample game and the graphic was not being used in any
manner within the project. My bet is that developers may come
with more clever ways of communicating screen orientation
issues to their users. Providing these routines are beyond
the scope of this project now.