Tasks refactoring and code compilation changes
I have been planning these and the upcoming changes for quite
a while now to improve some aspects of the template, simplify
code structure and address some overdue issues.

The bundler helper module was replaced by a leaner version.
The former implementation used some cryptic logic involving a
closure variable and was only devised as a temporary solution,
so I could focus in other details of the sample game and
development environment.

To accommodate to this change, the way how game code is
compiled for distribution needed to be changed, such that the
Phaser build chosen for development is no longer concatenated
with game code in a single file, as done before, but copied as
a separate, minified script file.