Changed several aspects of the sample game code
These changes were to be delivered with the ones that changed
directory structure. They are intended to make the sample game
more friendly to newcomers, removing noise and reventing some
dumb decisions only made to accommodate my own needs.

The `SplashScreen` class has been removed. It was there as a
sample prefab, but it was a very ineffective one, since its
simpler to just use the Phaser API to do so. In its place a
sample `Logo` class was created. It makes the sample game more
aligned with other Phaser game samples found elsewhere now.

The code responsible for making the game wait the audio
decoding has been removed. I doubt it was useful to anyone, so
I decided to remove it anyway.

For last, a slight change in the `app.main` routine, so now it
enumerates game states using `Object.entries`, and some code
comments reviewed for wording and grammar choices too.